requirements.txt
python-dotenv
pillow
requests